Fix studio projection provider registration
